+++ packages/orq-rc/src/orq_ai_sdk/knowledge.py
@@ -3,6 +3,16 @@
 from .basesdk import BaseSDK
 from orq_ai_sdk import models, utils
 from orq_ai_sdk._hooks import HookContext
+from orq_ai_sdk.models import (
+    createchunkop as models_createchunkop,
+    createdatasourceop as models_createdatasourceop,
+    createknowledgeop as models_createknowledgeop,
+    listchunksop as models_listchunksop,
+    listdatasourcesop as models_listdatasourcesop,
+    searchknowledgeop as models_searchknowledgeop,
+    updatechunkop as models_updatechunkop,
+    updateknowledgeop as models_updateknowledgeop,
+)
 from orq_ai_sdk.types import OptionalNullable, UNSET
 from orq_ai_sdk.utils import get_security_from_env
 from orq_ai_sdk.utils.unmarshal_json_response import unmarshal_json_response
@@ -206,7 +216,10 @@ class Knowledge(BaseSDK):
         path: str,
         description: Optional[str] = None,
         retrieval_settings: Optional[
-            Union[models.RetrievalSettings, models.RetrievalSettingsTypedDict]
+            Union[
+                models_createknowledgeop.RetrievalSettings,
+                models_createknowledgeop.RetrievalSettingsTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -311,7 +324,10 @@ class Knowledge(BaseSDK):
         path: str,
         description: Optional[str] = None,
         retrieval_settings: Optional[
-            Union[models.RetrievalSettings, models.RetrievalSettingsTypedDict]
+            Union[
+                models_createknowledgeop.RetrievalSettings,
+                models_createknowledgeop.RetrievalSettingsTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -593,8 +609,8 @@ class Knowledge(BaseSDK):
         path: Optional[str] = None,
         retrieval_settings: Optional[
             Union[
-                models.UpdateKnowledgeRetrievalSettings,
-                models.UpdateKnowledgeRetrievalSettingsTypedDict,
+                models_updateknowledgeop.UpdateKnowledgeRetrievalSettings,
+                models_updateknowledgeop.UpdateKnowledgeRetrievalSettingsTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -708,8 +724,8 @@ class Knowledge(BaseSDK):
         path: Optional[str] = None,
         retrieval_settings: Optional[
             Union[
-                models.UpdateKnowledgeRetrievalSettings,
-                models.UpdateKnowledgeRetrievalSettingsTypedDict,
+                models_updateknowledgeop.UpdateKnowledgeRetrievalSettings,
+                models_updateknowledgeop.UpdateKnowledgeRetrievalSettingsTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -993,21 +1009,29 @@ class Knowledge(BaseSDK):
         query: str,
         top_k: Optional[int] = None,
         threshold: Optional[float] = None,
-        search_type: Optional[models.SearchType] = "hybrid_search",
-        filter_by: Optional[Union[models.FilterBy, models.FilterByTypedDict]] = None,
+        search_type: Optional[models_searchknowledgeop.SearchType] = "hybrid_search",
+        filter_by: Optional[
+            Union[
+                models_searchknowledgeop.FilterBy,
+                models_searchknowledgeop.FilterByTypedDict,
+            ]
+        ] = None,
         search_options: Optional[
-            Union[models.SearchOptions, models.SearchOptionsTypedDict]
+            Union[
+                models_searchknowledgeop.SearchOptions,
+                models_searchknowledgeop.SearchOptionsTypedDict,
+            ]
         ] = None,
         rerank_config: Optional[
             Union[
-                models.SearchKnowledgeRerankConfig,
-                models.SearchKnowledgeRerankConfigTypedDict,
+                models_searchknowledgeop.SearchKnowledgeRerankConfig,
+                models_searchknowledgeop.SearchKnowledgeRerankConfigTypedDict,
             ]
         ] = None,
         agentic_rag_config: Optional[
             Union[
-                models.SearchKnowledgeAgenticRagConfig,
-                models.SearchKnowledgeAgenticRagConfigTypedDict,
+                models_searchknowledgeop.SearchKnowledgeAgenticRagConfig,
+                models_searchknowledgeop.SearchKnowledgeAgenticRagConfigTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -1134,21 +1158,29 @@ class Knowledge(BaseSDK):
         query: str,
         top_k: Optional[int] = None,
         threshold: Optional[float] = None,
-        search_type: Optional[models.SearchType] = "hybrid_search",
-        filter_by: Optional[Union[models.FilterBy, models.FilterByTypedDict]] = None,
+        search_type: Optional[models_searchknowledgeop.SearchType] = "hybrid_search",
+        filter_by: Optional[
+            Union[
+                models_searchknowledgeop.FilterBy,
+                models_searchknowledgeop.FilterByTypedDict,
+            ]
+        ] = None,
         search_options: Optional[
-            Union[models.SearchOptions, models.SearchOptionsTypedDict]
+            Union[
+                models_searchknowledgeop.SearchOptions,
+                models_searchknowledgeop.SearchOptionsTypedDict,
+            ]
         ] = None,
         rerank_config: Optional[
             Union[
-                models.SearchKnowledgeRerankConfig,
-                models.SearchKnowledgeRerankConfigTypedDict,
+                models_searchknowledgeop.SearchKnowledgeRerankConfig,
+                models_searchknowledgeop.SearchKnowledgeRerankConfigTypedDict,
             ]
         ] = None,
         agentic_rag_config: Optional[
             Union[
-                models.SearchKnowledgeAgenticRagConfig,
-                models.SearchKnowledgeAgenticRagConfigTypedDict,
+                models_searchknowledgeop.SearchKnowledgeAgenticRagConfig,
+                models_searchknowledgeop.SearchKnowledgeAgenticRagConfigTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -1277,7 +1309,10 @@ class Knowledge(BaseSDK):
         ending_before: Optional[str] = None,
         q: Optional[str] = None,
         status: Optional[
-            Union[models.QueryParamStatus, models.QueryParamStatusTypedDict]
+            Union[
+                models_listdatasourcesop.QueryParamStatus,
+                models_listdatasourcesop.QueryParamStatusTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -1380,7 +1415,10 @@ class Knowledge(BaseSDK):
         ending_before: Optional[str] = None,
         q: Optional[str] = None,
         status: Optional[
-            Union[models.QueryParamStatus, models.QueryParamStatusTypedDict]
+            Union[
+                models_listdatasourcesop.QueryParamStatus,
+                models_listdatasourcesop.QueryParamStatusTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -1481,7 +1519,10 @@ class Knowledge(BaseSDK):
         display_name: Optional[str] = None,
         file_id: Optional[str] = None,
         chunking_options: Optional[
-            Union[models.ChunkingOptions, models.ChunkingOptionsTypedDict]
+            Union[
+                models_createdatasourceop.ChunkingOptions,
+                models_createdatasourceop.ChunkingOptionsTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -1589,7 +1630,10 @@ class Knowledge(BaseSDK):
         display_name: Optional[str] = None,
         file_id: Optional[str] = None,
         chunking_options: Optional[
-            Union[models.ChunkingOptions, models.ChunkingOptionsTypedDict]
+            Union[
+                models_createdatasourceop.ChunkingOptions,
+                models_createdatasourceop.ChunkingOptionsTypedDict,
+            ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -2255,8 +2299,8 @@ class Knowledge(BaseSDK):
         datasource_id: str,
         request_body: Optional[
             Union[
-                List[models.CreateChunkRequestBody],
-                List[models.CreateChunkRequestBodyTypedDict],
+                List[models_createchunkop.CreateChunkRequestBody],
+                List[models_createchunkop.CreateChunkRequestBodyTypedDict],
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -2361,8 +2405,8 @@ class Knowledge(BaseSDK):
         datasource_id: str,
         request_body: Optional[
             Union[
-                List[models.CreateChunkRequestBody],
-                List[models.CreateChunkRequestBodyTypedDict],
+                List[models_createchunkop.CreateChunkRequestBody],
+                List[models_createchunkop.CreateChunkRequestBodyTypedDict],
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -2471,8 +2515,8 @@ class Knowledge(BaseSDK):
         q: Optional[str] = None,
         status: Optional[
             Union[
-                models.ListChunksQueryParamStatus,
-                models.ListChunksQueryParamStatusTypedDict,
+                models_listchunksop.ListChunksQueryParamStatus,
+                models_listchunksop.ListChunksQueryParamStatusTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -2580,8 +2624,8 @@ class Knowledge(BaseSDK):
         q: Optional[str] = None,
         status: Optional[
             Union[
-                models.ListChunksQueryParamStatus,
-                models.ListChunksQueryParamStatusTypedDict,
+                models_listchunksop.ListChunksQueryParamStatus,
+                models_listchunksop.ListChunksQueryParamStatusTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -3330,8 +3374,8 @@ class Knowledge(BaseSDK):
         embedding: Optional[List[float]] = None,
         metadata: Optional[
             Union[
-                Dict[str, models.UpdateChunkMetadata],
-                Dict[str, models.UpdateChunkMetadataTypedDict],
+                Dict[str, models_updatechunkop.UpdateChunkMetadata],
+                Dict[str, models_updatechunkop.UpdateChunkMetadataTypedDict],
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -3445,8 +3489,8 @@ class Knowledge(BaseSDK):
         embedding: Optional[List[float]] = None,
         metadata: Optional[
             Union[
-                Dict[str, models.UpdateChunkMetadata],
-                Dict[str, models.UpdateChunkMetadataTypedDict],
+                Dict[str, models_updatechunkop.UpdateChunkMetadata],
+                Dict[str, models_updatechunkop.UpdateChunkMetadataTypedDict],
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,