+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K7C40698X00J46WAN8Y2VF3K"
+        "01K7EMQCMCJ4K3ZPVWG785P8E1"
     )
     r"""The id of the resource"""