+++ packages/orq-rc/src/orq_ai_sdk/models/getbudgetop.py
@@ -154,5 +154,5 @@ class GetBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-12T11:48:13.882Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-13T11:19:00.603Z")
     r"""The date and time the resource was last updated"""