+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K7C40697WQWBEB67EB9H8D7X"
+        "01K7EMQCMB93KQZPD28B57EHZS"
     )
     r"""The id of the resource"""