+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K7C406995P33NV5RBHPP6RR0"
+        "01K7EMQCMDVHAC3K88WR7VGRRX"
     )
     r"""The id of the resource"""