+++ packages/orq-rc/RELEASES.md
@@ -3259,3 +3259,13 @@ Based on:
 - [python v3.14.0-rc.21] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.21] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.21 - packages/orq-rc
+
+## 2025-10-13 11:20:35
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.3 (2.723.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.22] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.22] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.22 - packages/orq-rc